implement answer count and complete comment exception disposal

+++ src/main/java/com/wenmrong/community1/community/controller/PublishController.java
@@ -25,7 +25,7 @@ public class PublishController {
 
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id")Integer id, Model model){
+    public String edit(@PathVariable(name = "id")Long id, Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title",question.getTitle());
         model.addAttribute("description",question.getDescription());
@@ -42,7 +42,7 @@ public class PublishController {
             @RequestParam(value = "title") String title,
             @RequestParam(value = "description") String description,
             @RequestParam(value = "tag") String tag,
-            @RequestParam(value = "id") Integer id,
+            @RequestParam(value = "id") Long id,
             HttpServletRequest request,
             Model model) {
             model.addAttribute("title",title);
